Fix for restart on standby Machine when registering AD-UserStore

+++ Modules/ArcGIS/DSCResources/ArcGIS_Portal/ArcGIS_Portal.psm1
@@ -644,6 +644,8 @@ function Set-TargetResource {
                 Write-Verbose $PortalSelfResponse
             }
         }
+        $StandByRestartRequired = $False
+        $StandByMachineName = ""
         if(-not($Join)){
             # set system property userstoreconfig > AD only done on primary machine. 
             #Dont need to restart standby machine since join hasn't happened yet.
@@ -681,11 +683,21 @@ function Set-TargetResource {
                     $RestartRequired = $true
                 }
             }
+            if($RestartRequired){
+                $machines = Get-RegisteredMachine -PortalHostName $FQDN -Token $token.token
+                if($machines.machines.count > 2){
+                    $StandByRestartRequired = $true
+                    $StandByMachineName = ($machine.machines| Where-Object { (Get-FQDN $_.machineName) -ne $FQDN }).machineName
+                }
+            }
         }
         if ($RestartRequired) 
         {
             Restart-PortalService
             Wait-ForUrl "https://$($FQDN):7443/arcgis/portaladmin" -HttpMethod 'GET'
+            if($StandByRestartRequired){
+                Restart-PortalService -MachineName $StandByMachineName
+            }
         }
 
     }
@@ -957,15 +969,25 @@ function Restart-PortalService {
     [OutputType([System.Boolean])]
     param
     (
+        [Parameter(Mandatory=$false)]    
         [System.String]
-        $ServiceName = 'Portal for ArcGIS'
+        $ServiceName = 'Portal for ArcGIS',
+        
+        [Parameter(Mandatory=$false)]
+        [System.String]
+        $MachineName
     )
 
     try {
         Write-Verbose "Restarting Service $ServiceName"
-        Stop-Service -Name $ServiceName -Force -ErrorAction Ignore
+        if($MachineName -ne $null -or $MachineName -ne ""){
+            Get-Service -Name $ServiceName -ComputerName $MachineName | Set-Service -Status Stopped -Force -ErrorAction Ignore
+        }else{
+            Stop-Service -Name $ServiceName -Force -ErrorAction Ignore
+        }
+        
         Write-Verbose 'Stopping the service' 
-        Wait-ForServiceToReachDesiredState -ServiceName $ServiceName -DesiredState 'Stopped'
+        Wait-ForServiceToReachDesiredState -ServiceName $ServiceName -DesiredState 'Stopped' -MachineName $MachineName
         Write-Verbose 'Stopped the service'
     }
     catch {
@@ -974,8 +996,13 @@ function Restart-PortalService {
 
     try {
         Write-Verbose 'Starting the service'
-        Start-Service -Name $ServiceName -ErrorAction Ignore        
-        Wait-ForServiceToReachDesiredState -ServiceName $ServiceName -DesiredState 'Running'
+        if($MachineName -ne $null -or $MachineName -ne ""){
+            Get-Service -Name $ServiceName -ComputerName $MachineName | Set-Service -Status Running -Force -ErrorAction Ignore
+        }else{
+            Start-Service -Name $ServiceName -ErrorAction Ignore
+        }
+
+        Wait-ForServiceToReachDesiredState -ServiceName $ServiceName -DesiredState 'Running' -MachineName $MachineName
         Write-Verbose "Restarted Service '$ServiceName'"
     }
     catch {
@@ -1029,6 +1056,29 @@ function Set-LoggingLevel {
         }
     }
     $ServiceRestartRequired
+}
+
+function Get-RegisteredMachine {
+    [CmdletBinding()]
+    param(
+        [System.String]
+        $PortalHostName = 'localhost',
+
+        [System.String]
+        $SiteName = 'arcgis',
+
+        [System.Int32]
+        $Port = 7443,
+
+        [System.String]
+        $Token,
+
+        [System.String]
+        $Referer = 'http://localhost'
+    )   
+
+    Invoke-ArcGISWebRequest -Url ("https://$($PortalHostName):$($Port)/$SiteName/portaladmin/machines") `
+                        -HttpFormParameters @{ f = 'json'; token = $Token; } -Referer $Referer -HttpMethod 'GET'
 }
 
 function Get-PortalSecurityConfig {
